+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added API to render specific settings during a frame
 - Added stadia to the supported platforms (2019.3)
 - Enabled cascade blends settings in the HD Shadow component
+- Added Hardware Dynamic Resolution support. 
 
 ### Fixed
 - Fixed deserialization crash at runtime